requirements.txt
# KRX Auto Crawling Dependencies

# Core
setuptools>=65.0.0

# Web scraping
cloudscraper>=1.2.71
beautifulsoup4>=4.12.0
playwright>=1.40.0

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0


# Async
asyncio>=3.4.3

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
